Safari-611.1.10/LayoutTests/fast/dom/StyleSheet/get-stylesheet-byname.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This test verifies that a StyleSheet object will be returned instead of a HTMLStyleElement when calling document.styleSheets named property getter.");

var styleElement = document.createElement("style");
styleElement.setAttribute("id", "test");
document.head.appendChild(styleElement);
shouldBe('document.styleSheets["test"]', 'styleElement.sheet');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
